ovirtlago: move OvirtPrefix from __init__.py to prefix.py
This patch moves the OvirtPrefix class to  prefix.py,
leaving an empty ovirtlago/__init__.py file.

When OvirtPrefix is defined in __init__.py, it is declared whenever
the 'ovirtlago' package is imported(even if importing different
classes, i.e.: from ovirtlago import virt), this has 2 problems:

1. (Theoretical) - there is no demand for this coupling.
2. It requires installing all 'OvirtPrefix' imports in the unit testing
environment, even if only a sub-module is tested.
