+++ ovirtlago/testlib.py
@@ -41,7 +41,7 @@ def get_test_prefix():
     global _test_prefix
     if _test_prefix is None:
         cur_workdir_path = os.environ.get('LAGO_WORKDIR_PATH', os.curdir)
-        workdir = ovirtlago.OvirtWorkdir(path=cur_workdir_path)
+        workdir = ovirtlago.prefix.OvirtWorkdir(path=cur_workdir_path)
         _test_prefix = workdir.get_prefix('current')
     return _test_prefix
